# feat(common): add `append_one_row_from_array_elements` to `DataChunkBuilder`