I hereby agree to the terms of the [Singularity Data, Inc. Contributor License Agreement](https://gist.github.com/skyzh/0663682a70b0edde7ae991492f2314cb#file-s9y_cla).

## What's changed and what's your intention?

When doing batch join, there are cases where multiple rows should be appended to the data chunk builder and they are usually combinations of rows from two tables. Currently, we handle this by chaining the values of two sub-rows and calling `append_one_row_from_datum_refs` multiple times, for example (from https://github.com/singularity-data/risingwave/pull/4017):
```rust
#[for_await]
for probe_chunk in probe_side.execute() {
    let probe_chunk = probe_chunk?;
    let probe_keys = K::build(&probe_key_idxs, &probe_chunk)?;
    for (probe_row, probe_key) in probe_chunk.rows().zip_eq(probe_keys) {
        let mut matched_build_row_id = hash_map.get(&probe_key);
        while let Some(&build_row_id) = matched_build_row_id {
            let build_chunk = &build_side[build_row_id.chunk_id()];
            let build_row = build_chunk.row_at_unchecked_vis(build_row_id.row_id());
            let datum_refs = probe_row.values().chain(build_row.values());
            if let Some(spilled) =
                chunk_builder.append_one_row_from_datum_refs(datum_refs)?
            {
                yield spilled
            }
            matched_build_row_id = next_row_with_same_key[build_row_id].as_ref();
        }
    }
}
```

To handle this case more efficiently, a new method `append_one_row_from_array_elements` can be added to `DataChunkBuilder`. Example usage:
```rust
#[for_await]
for probe_chunk in probe_side.execute() {
    let probe_chunk = probe_chunk?;
    let probe_keys = K::build(&probe_key_idxs, &probe_chunk)?;
    for (probe_row_id, probe_key) in probe_keys.iter().enumerate() {
        let mut matched_build_row_id = hash_map.get(&probe_key);
        while let Some(&build_row_id) = matched_build_row_id {
            let build_chunk = &build_side[build_row_id.chunk_id()];
            if let Some(chunk) = chunk_builder.append_one_row_from_array_elements(
                probe_chunk.columns().iter().map(|c| c.array_ref()),
                probe_row_id,
                build_chunk.columns().iter().map(|c| c.array_ref()),
                build_row_id.row_id(),
            )? {
                yield chunk
            }
            matched_build_row_id = next_row_with_same_key[build_row_id].as_ref();
        }
    }
}
```

## Benchmark Report (WIP)

| Type | Before | After |
| --- | --- | --- |
| HashJoinInner(32) | 479.92 µs | **315.48 µs** |
| HashJoinInner(128) | 473.44 µs | **300.19 µs** |
| HashJoinInner(512) | 439.82 µs | **290.78 µs** |

## Checklist

- [x] I have written necessary rustdoc comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Refer to a related PR or issue link (optional)
https://github.com/singularity-data/risingwave/pull/4017